+++ Changelog.md
@@ -1,3 +1,8 @@
+# 1.2.1-Beta
+ - Logic fixes with certain collectibles
+   - Inferno honeycomb and Cheato page
+   - Pig Pool jiggy
+
 # 1.2-Beta
 - 2 new Goals added:
   - Minigames Goal: Complete all minigames with optional Canary Mary 4.